agent-api/pyproject.toml
[project]
name = "agent-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "fastmcp>=2.6.0",
    "groq>=0.26.0",
    "instructor>=1.8.3",
    "ipykernel>=6.29.5",
    "loguru>=0.7.3",
    "opik>=1.7.29",
    "pixeltable>=0.3.15",
    "pydantic-settings>=2.9.1",
    "ruff>=0.11.12",
]
